% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_htmlwidgets}
\alias{data_htmlwidgets}
\title{Sample dataset for a gallery}
\format{A data frame with six rows and four variables:
\describe{
\item{href}{\code{character}, URL target for the link}
\item{image}{\code{character}, URL for the image}
\item{title}{\code{character}, text for the title}
\item{subtitle}{\code{character} text for the subtitle}
}}
\source{
\url{http://www.diamondse.info/}
}
\usage{
data_htmlwidgets
}
\description{
A dataset containg the information needed to build
a gallerywidget.
}
\keyword{datasets}
