% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gallerywidget.R
\name{gallerywidget}
\alias{gallerywidget}
\title{Create a gallerywidget}
\usage{
gallerywidget(data)
}
\arguments{
\item{data}{\code{data.frame} with columns \code{href}, \code{image}, \code{title}, \code{subtitle},
see \strong{Details}}
}
\value{
object returned by \code{\link[htmlwidgets:createWidget]{htmlwidgets::createWidget()}}
}
\description{
Create a gallerywidget
}
\examples{
  gallerywidget(data_htmlwidgets)
}
